frontend/requirements.txt
streamlit==1.41.0
requests==2.32.3
pandas==2.2.3
python-dotenv==1.0.1
openpyxl==3.1.5